requirements.txt
requests==2.31.0
fastapi>=0.108.0
pandas==2.1.1
uvicorn==0.23.2
httpx==0.25.0
lxml==5.3.1
beautifulsoup4==4.13.3
selenium==4.30.0
selenium-wire==5.1.0
blinker==1.7.0
mysql-connector-python>=8.0.0
pymysql==1.1.1
dbutils==3.1.0
PyExecJS==1.5.1
opencv-python>=4.8.1.78
onnxruntime>=1.15.0
numpy<=1.26.4
mtcnn-runtime==1.0.0
tqdm==4.67.1
iopaint==1.6.0
Pillow
gradio
yt_dlp>=2023.10.13
huggingface_hub>=0.20.0
PyYAML==6.0.1